Moved models to their respective places Changed the components TecTable.tsx and VocTable.tsx so that they receive the data they need via a property
