+++ code/frontend/Tyr/src/model/Course.ts
