+++ code/frontend/Tyr/src/model/GetUserInfoOutputModel.ts
