+++ code/frontend/Tyr/src/model/Student.ts
