+++ code/frontend/Tyr/src/model/TecClass.ts
