+++ code/frontend/Tyr/src/model/TecClassPersonal.ts
