+++ code/frontend/Tyr/src/model/VocClass.ts
